refactor(home): clean up home page code

+++ src/main/java/com/papercraft/model/Product.java
@@ -2,7 +2,7 @@ package com.papercraft.model;
 
 import java.io.Serializable;
 import java.math.BigDecimal;
-import java.security.Timestamp;
+import java.sql.Timestamp;
 import java.util.List;
 
 public class Product implements Serializable {
@@ -13,12 +13,11 @@ public class Product implements Serializable {
     private String productDescription;
     private String productDetail;
     private String brand;
-    private BigDecimal price;
-    private BigDecimal originPrice;
-    private BigDecimal discount;
+    private double price;
+    private double originPrice;
+    private double discount;
     private int stockQuantity;
     private Timestamp createdAt;
-
 
     private Integer quantity;
     private String type;
@@ -26,9 +25,10 @@ public class Product implements Serializable {
     private List<String> imageList;
     private BigDecimal avgRating;
 
-    public Product() {}
+    public Product() {
+    }
 
-    public Product(int id, int categoryId, String productName, String descriptionThumbnail, String productDescription, String productDetail, String brand, BigDecimal price, BigDecimal originPrice, BigDecimal discount, int stockQuantity, Timestamp createdAt, Integer quantity, String type, String thumbnail, List<String> imageList, BigDecimal avgRating) {
+    public Product(int id, int categoryId, String productName, String descriptionThumbnail, String productDescription, String productDetail, String brand, double price, double originPrice, double discount, int stockQuantity, Timestamp createdAt, Integer quantity, String type, String thumbnail, List<String> imageList, BigDecimal avgRating) {
         this.id = id;
         this.categoryId = categoryId;
         this.productName = productName;
@@ -108,7 +108,7 @@ public class Product implements Serializable {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 
@@ -116,7 +116,7 @@ public class Product implements Serializable {
         return originPrice;
     }
 
-    public void setOriginPrice(BigDecimal originPrice) {
+    public void setOriginPrice(double originPrice) {
         this.originPrice = originPrice;
     }
 
@@ -124,7 +124,7 @@ public class Product implements Serializable {
         return discount;
     }
 
-    public void setDiscount(BigDecimal discount) {
+    public void setDiscount(double discount) {
         this.discount = discount;
     }
 
@@ -184,8 +184,7 @@ public class Product implements Serializable {
         this.avgRating = avgRating;
     }
 
-    public void quantityUp(int qty){
+    public void quantityUp(int qty) {
         this.quantity += qty;
-
     }
 }
